data_analysis_agent/requirements.txt
openai
duckdb
pandas
numpy
phidata
streamlit
